+++ src/components/Header/Header.tsx
@@ -13,35 +13,35 @@ export const Header: FC<HeaderProps> = () => {
   return (
     <header className={style.header_wrapper}>
       <div className={style.header_content}>
-        <Link href='/' className={style.header_logo_box}>
+        <Link href="/" className={style.header_logo_box}>
           <div className={style.header_logo_wrapper}>
-            <img src='/images/header_logo.png' alt='Logo' />
+            <img src="/images/header_logo.png" alt="Logo" />
           </div>
           <h1 className={style.header_logo_text}>TESTTASK</h1>
         </Link>
         <div className={style.header_nav_wrapper}>
           <div className={style.header_btn_wrapper}>
-            <Button href='/#teamMembers' className={style.button}>
+            <Button href="/#teamMembers" className={style.button}>
               Users
             </Button>
           </div>
           {user ? (
             <div className={style.header_avatar_wrapper}>
               <img
-                src={user.photoURL || '/images/default_avatar.png'} // Заміна на аватар користувача
-                alt='User Avatar'
+                src={user.photoURL || "./public/images/avatar"} // Заміна на аватар користувача
+                alt="User Avatar"
                 className={style.user_avatar}
               />
             </div>
           ) : (
             <>
               <div className={style.header_btn_wrapper}>
-                <Button href='/#signUpForm' className={style.button}>
+                <Button href="/#signUpForm" className={style.button}>
                   Sign up
                 </Button>
               </div>
               <div className={style.header_btn_wrapper}>
-                <Link className={style.button} href='/login'>
+                <Link className={style.button} href="/login">
                   Login
                 </Link>
               </div>
