+++ src/hooks/useImageValidation.ts
