+++ src/pages/login.tsx
@@ -1,14 +1,10 @@
-import { Input } from "@/components/Form/components/index";
-import { LoginForm } from "@/modules/Login_component";
-import style from "./login.module.scss"
-import Link from "next/link";
-import { ChangeEvent, FocusEvent } from "react";
-
+import { LoginForm } from "@/components/Form/Login-Form/Login_component";
+import style from "./login.module.scss";
 
 export default function Login() {
-    return (
-      <div className={style.login_wrapper}>
-        <LoginForm />
-      </div>
-    );
-};
+  return (
+    <div className={style.login_wrapper}>
+      <LoginForm />
+    </div>
+  );
+}
